[integer-ext] Add methods to Integer class
Extending functionality to the Integer class.

positive? : returns true if and only if the integer is greater than 0.
nonpositive? : returns true if and only if the integer is not positive.
negative? : returns true if and only if the integer is less than 0.
nonnegative? : returns true if and only if the integer is not negative.