Add private method to enhance readability
module Goby contains a private method `run_turn` that had a check at the
end of the function definition. While it is acceptable practice to use a
not operator ("!") before a function call, it can decrease the
readability of logic and increases cognitive overhead to understand the
meaning of the code. Creating a new private method `continue_game?`
makes explicit the purpose of that logic check at the end of the
`run_turn` method within the context of the game.